+++ Saule/Serialization/ResourceDeserializer.cs
@@ -108,7 +108,7 @@ namespace Saule.Serialization
             foreach (var rel in child["relationships"] ?? new JArray())
             {
                 var prop = rel as JProperty;
-                result.Add(prop?.Name.ToPascalCase(), ToFlatStructure(prop?.Value));
+                result.Add(_propertyNameConverter.ToModelPropertyName(prop?.Name), ToFlatStructure(prop?.Value));
             }
 
             return result;