+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Master (upcoming 1.8)
 
+- [**BUGFIX**] Support camel case serialization for relationship names
+
 ## Version 1.7.1
 
 - [**BUGFIX**] Support camel case serialization for nested properties in attributes and meta hashes